http proxy: extract http transport to separate file
* NewHTTPProxy accepts t *http.Transport instead of r *net.Resolver
* The configuration is standalone and can be used outside http proxy context
