+++ http_proxy.go
@@ -10,86 +10,11 @@ import (
 	"net"
 	"net/http"
 	"net/url"
-	"runtime"
-	"time"
 
 	"github.com/elazarl/goproxy"
 	"github.com/saucelabs/forwarder/middleware"
 	"github.com/saucelabs/pacman"
 )
-
-type HTTPTransportConfig struct {
-	// DialTimeout is the maximum amount of time a dial will wait for
-	// a connect to complete.
-	//
-	// With or without a timeout, the operating system may impose
-	// its own earlier timeout. For instance, TCP timeouts are
-	// often around 3 minutes.
-	DialTimeout time.Duration `json:"dial_timeout"`
-
-	// KeepAlive specifies the interval between keep-alive
-	// probes for an active network connection.
-	// If zero, keep-alive probes are sent with a default value
-	// (currently 15 seconds), if supported by the protocol and operating
-	// system. Network protocols or operating systems that do
-	// not support keep-alives ignore this field.
-	// If negative, keep-alive probes are disabled.
-	KeepAlive time.Duration `json:"keep_alive"`
-
-	// TLSHandshakeTimeout specifies the maximum amount of time waiting to
-	// wait for a TLS handshake. Zero means no timeout.
-	TLSHandshakeTimeout time.Duration `json:"tls_handshake_timeout"`
-
-	// MaxIdleConns controls the maximum number of idle (keep-alive)
-	// connections across all hosts. Zero means no limit.
-	MaxIdleConns int `json:"max_idle_conns"`
-
-	// MaxIdleConnsPerHost, if non-zero, controls the maximum idle
-	// (keep-alive) connections to keep per-host. If zero,
-	// DefaultMaxIdleConnsPerHost is used.
-	MaxIdleConnsPerHost int `json:"max_idle_conns_per_host"`
-
-	// MaxConnsPerHost optionally limits the total number of
-	// connections per host, including connections in the dialing,
-	// active, and idle states. On limit violation, dials will block.
-	//
-	// Zero means no limit.
-	MaxConnsPerHost int `json:"max_conns_per_host"`
-
-	// IdleConnTimeout is the maximum amount of time an idle
-	// (keep-alive) connection will remain idle before closing
-	// itself.
-	// Zero means no limit.
-	IdleConnTimeout time.Duration `json:"idle_conn_timeout"`
-
-	// ResponseHeaderTimeout, if non-zero, specifies the amount of
-	// time to wait for a server's response headers after fully
-	// writing the request (including its body, if any). This
-	// time does not include the time to read the response body.
-	ResponseHeaderTimeout time.Duration `json:"response_header_timeout"`
-
-	// ExpectContinueTimeout, if non-zero, specifies the amount of
-	// time to wait for a server's first response headers after fully
-	// writing the request headers if the request has an
-	// "Expect: 100-continue" header. Zero means no timeout and
-	// causes the body to be sent immediately, without
-	// waiting for the server to approve.
-	// This time does not include the time to send the request header.
-	ExpectContinueTimeout time.Duration `json:"expect_continue_timeout"`
-}
-
-func DefaultHTTPTransportConfig() *HTTPTransportConfig {
-	// The default values are taken from [hashicorp/go-cleanhttp](https://github.com/hashicorp/go-cleanhttp/blob/a0807dd79fc1680a7b1f2d5a2081d92567aab97d/cleanhttp.go#L19.
-	return &HTTPTransportConfig{
-		DialTimeout:           30 * time.Second,
-		KeepAlive:             30 * time.Second,
-		MaxIdleConns:          100,
-		IdleConnTimeout:       90 * time.Second,
-		TLSHandshakeTimeout:   10 * time.Second,
-		ExpectContinueTimeout: 1 * time.Second,
-		MaxIdleConnsPerHost:   runtime.GOMAXPROCS(0) + 1,
-	}
-}
 
 type HTTPProxyConfig struct {
 	// UpstreamProxy is the upstream proxy , ex. http://user:password@127.0.0.1:8080.
@@ -120,16 +45,10 @@ type HTTPProxyConfig struct {
 	// - usr3:pwd3@bar.foo:8080
 	// HTTPProxy will add basic auth headers for requests to these URLs.
 	SiteCredentials []string `json:"site_credentials"`
-
-	// Transport specifies http.Transport configuration used when making Transport requests.
-	// If nil, DefaultHTTPTransportConfig will be used.
-	Transport *HTTPTransportConfig `json:"http"`
 }
 
 func DefaultHTTPProxyConfig() *HTTPProxyConfig {
-	return &HTTPProxyConfig{
-		Transport: DefaultHTTPTransportConfig(),
-	}
+	return &HTTPProxyConfig{}
 }
 
 func (c *HTTPProxyConfig) Validate() error {
@@ -161,13 +80,15 @@ type HTTPProxy struct {
 	log       Logger
 }
 
-func NewHTTPProxy(cfg *HTTPProxyConfig, r *net.Resolver, log Logger) (*HTTPProxy, error) {
-	if cfg.Transport == nil {
-		cfg.Transport = DefaultHTTPTransportConfig()
-	}
-
+func NewHTTPProxy(cfg *HTTPProxyConfig, t *http.Transport, log Logger) (*HTTPProxy, error) {
 	if err := cfg.Validate(); err != nil {
 		return nil, err
+	}
+
+	// If not set, use http.DefaultTransport.
+	if t == nil {
+		log.Infof("HTTP transport not configured, using standard library default")
+		t = http.DefaultTransport.(*http.Transport).Clone() //nolint:forcetypeassert // we know it's a *http.Transport
 	}
 
 	// Parse site credential list into map of host:port -> base64 encoded input.
@@ -178,11 +99,11 @@ func NewHTTPProxy(cfg *HTTPProxyConfig, r *net.Resolver, log Logger) (*HTTPProxy
 
 	p := &HTTPProxy{
 		config:    *cfg,
+		transport: t,
 		userInfo:  m,
 		basicAuth: middleware.NewProxyBasicAuth(),
 		log:       log,
 	}
-	p.configureTransport(r)
 
 	if p.config.PAC != nil {
 		pacParser, err := pacman.New(p.config.PAC.String(), p.config.PACProxiesCredentials...)
@@ -195,28 +116,6 @@ func NewHTTPProxy(cfg *HTTPProxyConfig, r *net.Resolver, log Logger) (*HTTPProxy
 	p.configureProxy()
 
 	return p, nil
-}
-
-func (hp *HTTPProxy) configureTransport(r *net.Resolver) {
-	hp.log.Infof("Using HTTP transport config: %+v", *hp.config.Transport)
-
-	c := hp.config.Transport
-	hp.transport = &http.Transport{
-		Proxy: nil,
-		DialContext: defaultTransportDialContext(&net.Dialer{
-			Timeout:   c.DialTimeout,
-			KeepAlive: c.KeepAlive,
-			Resolver:  r,
-		}),
-		MaxIdleConns:          c.MaxIdleConns,
-		IdleConnTimeout:       c.IdleConnTimeout,
-		TLSHandshakeTimeout:   c.TLSHandshakeTimeout,
-		ExpectContinueTimeout: c.ExpectContinueTimeout,
-		ForceAttemptHTTP2:     true,
-		MaxIdleConnsPerHost:   c.MaxIdleConnsPerHost,
-	}
-
-	hp.transport.Proxy = nil
 }
 
 func defaultTransportDialContext(dialer *net.Dialer) func(context.Context, string, string) (net.Conn, error) {
